Move MailViewRouterDelegate out of home.dart and into its own file
